Some geometry utilities for converting between bounding boxes and limits

+++ functions/geometry/sz2bb.m
@@ -2,7 +2,7 @@ function bb = sz2bb(sz)
 %SZ2BB Converts the output of size to a bounding box around an image of the same size.
 %   Produces the same output as ref_bb(imref2d(sz)).
 
-bb = minaabb([0.5 0.5; sz + [0.5 0.5]]);
+bb = minaabb(double([0.5 0.5; sz + [0.5 0.5]]));
 
 end
